Moved RAII classes to header
