+++ tools/clang/unittests/HLSLHost/HLSLHost.cpp
@@ -13,6 +13,7 @@
 #include "dxc/Support/Global.h"
 #include "dxc/Support/WinIncludes.h"
 #include "dxc/Support/microcom.h"
+#include "dxc/Support/RAIIClasses.h" // RAII Helpers
 #include "llvm/Support/Compiler.h" // for LLVM_FALLTHROUGH
 #include <algorithm>
 #include <atlcoll.h>
@@ -37,88 +38,6 @@ Pending work for rendering hosting.
 
 // Forward declarations.
 class ServerFactory;
-
-// RAII helpers.
-class HhEvent {
-public:
-  HANDLE m_handle = INVALID_HANDLE_VALUE;
-
-public:
-  HRESULT Init() {
-    if (m_handle == INVALID_HANDLE_VALUE) {
-      m_handle = CreateEvent(nullptr, TRUE, FALSE, nullptr);
-      if (m_handle == INVALID_HANDLE_VALUE) {
-        return HRESULT_FROM_WIN32(GetLastError());
-      }
-    }
-    return S_OK;
-  }
-  void SetEvent() { ::SetEvent(m_handle); }
-  void ResetEvent() { ::ResetEvent(m_handle); }
-  ~HhEvent() {
-    if (m_handle != INVALID_HANDLE_VALUE) {
-      CloseHandle(m_handle);
-    }
-  }
-};
-
-class HhCriticalSection {
-private:
-  CRITICAL_SECTION m_cs;
-
-public:
-  HhCriticalSection() { InitializeCriticalSection(&m_cs); }
-  ~HhCriticalSection() { DeleteCriticalSection(&m_cs); }
-  class Lock {
-  private:
-    CRITICAL_SECTION *m_pLock;
-
-  public:
-    Lock() = delete;
-    Lock(const Lock &) = delete;
-    Lock(Lock &&other) { std::swap(m_pLock, other.m_pLock); }
-    Lock(CRITICAL_SECTION *pLock) : m_pLock(pLock) {
-      EnterCriticalSection(m_pLock);
-    }
-    ~Lock() {
-      if (m_pLock)
-        LeaveCriticalSection(m_pLock);
-    }
-  };
-  Lock LockCS() { return Lock(&m_cs); }
-};
-
-class ClassObjectRegistration {
-private:
-  DWORD m_reg = 0;
-  HRESULT m_hr = E_FAIL;
-
-public:
-  HRESULT Register(REFCLSID rclsid, IUnknown *pUnk, DWORD dwClsContext,
-                   DWORD flags) {
-    m_hr = CoRegisterClassObject(rclsid, pUnk, dwClsContext, flags, &m_reg);
-    return m_hr;
-  }
-  ~ClassObjectRegistration() {
-    if (SUCCEEDED(m_hr))
-      CoRevokeClassObject(m_reg);
-  }
-};
-
-class CoInit {
-private:
-  HRESULT m_hr = E_FAIL;
-
-public:
-  HRESULT Initialize(DWORD dwCoInit) {
-    m_hr = CoInitializeEx(nullptr, dwCoInit);
-    return m_hr;
-  }
-  ~CoInit() {
-    if (SUCCEEDED(m_hr))
-      CoUninitialize();
-  }
-};
 
 // Globals.
 static const GUID
